Adds support for multi-value attributes
A recording form may ask a question such as 'What are the two main
colours of the ladybird?' and offer a list of options with checkboxes.
The warehouse supports this by allowing multiple rows in any
attribute_values table for an attribute_id.
By setting an array value on an attribute and posting that array, the
SDK can now use this feature.